Updated cache server
